refactor: concurrency model simplification
Replace use of event stream with a mutex wrapper around game
state. Uses Arc to enable reference counting on the Mutex
wrapper. Allows GTK callbacks to interact with the game instance
directly.

Required setting an unsafe Send declaration for Game. Maybe this is
bad, still a bit of a Rust newbie..